@@ -2,10 +2,10 @@
   "name": "@stock-charts/client",
   "description": "Angular frontend for stock charts application",
   "type": "module",
-  "packageManager": "pnpm@10.27.0",
+  "packageManager": "pnpm@10.29.1",
   "engines": {
-    "node": ">=24.12.0",
-    "pnpm": ">=10.27.0"
+    "node": ">=24.13.0",
+    "pnpm": ">=10.29.1"
   },
   "scripts": {
     "ng": "ng",
@@ -26,19 +26,19 @@
     "clean": "rm -rf .angular coverage dist node_modules test-results"
   },
   "dependencies": {
-    "@angular/animations": "21.0.6",
-    "@angular/cdk": "21.0.5",
-    "@angular/common": "21.0.6",
-    "@angular/compiler": "21.0.7",
-    "@angular/core": "21.0.8",
-    "@angular/forms": "21.0.6",
-    "@angular/material": "21.0.5",
-    "@angular/platform-browser": "21.0.6",
-    "@angular/platform-browser-dynamic": "21.0.6",
-    "@angular/router": "21.0.6",
-    "@angular/service-worker": "21.0.6",
+    "@angular/animations": "21.1.3",
+    "@angular/cdk": "21.1.3",
+    "@angular/common": "21.1.3",
+    "@angular/compiler": "21.1.3",
+    "@angular/core": "21.1.3",
+    "@angular/forms": "21.1.3",
+    "@angular/material": "21.1.3",
+    "@angular/platform-browser": "21.1.3",
+    "@angular/platform-browser-dynamic": "21.1.3",
+    "@angular/router": "21.1.3",
+    "@angular/service-worker": "21.1.3",
     "@ctrl/tinycolor": "4.2.0",
-    "@ng-matero/extensions": "21.1.2",
+    "@ng-matero/extensions": "21.1.3",
     "chart.js": "4.5.1",
     "chartjs-adapter-date-fns": "3.0.0",
     "chartjs-plugin-annotation": "3.1.0",
@@ -47,23 +47,23 @@
     "rxjs": "7.8.2",
     "tslib": "2.8.1",
     "uuid": "13.0.0",
-    "zone.js": "0.15.1"
+    "zone.js": "0.16.0"
   },
   "devDependencies": {
-    "@angular/build": "21.0.4",
-    "@angular/cli": "21.0.4",
-    "@angular/compiler-cli": "21.0.6",
+    "@angular/build": "21.1.3",
+    "@angular/cli": "21.1.3",
+    "@angular/compiler-cli": "21.1.3",
     "@eslint/js": "9.39.2",
-    "@types/node": "25.0.3",
-    "@vitest/coverage-v8": "4.0.16",
+    "@types/node": "25.2.1",
+    "@vitest/coverage-v8": "4.0.18",
     "@vitest/eslint-plugin": "1.6.6",
-    "angular-eslint": "21.1.0",
+    "angular-eslint": "21.2.0",
     "eslint": "9.39.2",
     "eslint-config-prettier": "10.1.8",
-    "eslint-plugin-prettier": "5.5.4",
-    "jsdom": "26.1.0",
+    "eslint-plugin-prettier": "5.5.5",
+    "jsdom": "28.0.0",
     "typescript": "5.9.3",
-    "typescript-eslint": "8.52.0",
-    "vitest": "4.0.16"
+    "typescript-eslint": "8.54.0",
+    "vitest": "4.0.18"
   }
 }
